requirements.txt
asyncio==3.4.3
atproto==0.0.56
feedparser==6.0.11
Mastodon.py==1.8.1
mautrix==0.20.7
MechanicalSoup==1.3.0
pillow==10.4.0
PyYAML==6.0.2
requests==2.32.3
telepot==12.7
tweepy==4.14.0
pythorhead==0.31.0